src/func/mcp-spotify/pyproject.toml
[tool.poetry]
name = "mcp-spotify"
version = "0.1.0"
description = "MCP Spotify Azure Function for music streaming integration"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
azure-functions = "^1.21.3"
requests = "^2.31.0"
azure-identity = "^1.16.0"
spotipy = "^2.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"